Renaming stats to valueblock. Issue #47
